requirements.txt
azure-storage >= 0.34.3